+++ R/proportions.R
@@ -6,9 +6,10 @@
 #' \code{\link[qdap]{termco}}).
 #' @param \ldots Arguments passed to proportions method of other classes.
 #' @export
-#' @seealso \code{\link[qdap]{scores}}
-#' @seealso \code{\link[qdap]{counts}}
-#' @seealso \code{\link[qdap]{preprocessed}}
+#' @seealso \code{\link[qdap]{scores}},
+#' \code{\link[qdap]{counts}},
+#' \code{\link[qdap]{preprocessed}},
+#' \code{\link[qdap]{visual}}
 #' @return Returns a data.frame of proportions.
 proportions <-
 function(x, ...){